backend/requirements.txt
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
Django==3.2
django-admin==2.0.1
django-cors-headers==3.14.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==23.1
django-six==1.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.0
frozenlist==1.3.3
future==0.18.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
netmiko==4.1.2
ntc-templates==3.3.0
paramiko==3.1.0
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyserial==3.5
pytz==2022.6
requests==2.28.1
rfc3986==1.5.0
scp==0.14.5
six==1.10.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
tenacity==8.2.2
textfsm==1.1.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
uvicorn==0.21.1
